refactor(trips): hoist exif/transform into server-image-safe modules
Move the pure, dependency-light EXIF and KML transform modules from
trips/backfill/ (excluded from the server image) up to trips/ so they
compile into the monolith server library for the upcoming live ingest
and SSR read paths. The heavy S3/CLI orchestration (backfill/main.py,
with boto3/typer/httpx) stays excluded.

- git mv exif.py, transform.py and their tests to trips/
- update imports to trips.exif / trips.transform
- the monolith_backend glob (trips/**/*.py minus trips/backfill/**) now
  ships the modules; add @pip//pillow and @pip//defusedxml to its deps
- rename trips_backfill_{exif,transform}_test -> trips_{exif,transform}_test,
  retargeting srcs and deps at :monolith_backend
