+++ projects/monolith/trips/exif.py
